First boss, boss drop, view equipment is functional
If you check Start_Fight(), you will see that the method specifically checks if what you just killed was a boss, and what boss it was, and then drops the appropriate loot as defined in Monster_Class. Monster_Class for the boss does NOT define a prefix, as it does not need one, and gives an adropconifg, UNLIKE conventional monsters. Works better this way because you can still get strong using skill points and racial attributes. Armor will be like a trophy.